+++ contracts/pair/src/contract.rs
@@ -22,6 +22,7 @@ pub struct LiquidityPool;
 pub trait LiquidityPoolTrait {
     // Sets the token contract addresses for this pool
     // token_wasm_hash is the WASM hash of the deployed token contract for the pool share token
+    #[allow(clippy::too_many_arguments)]
     fn initialize(
         env: Env,
         token_wasm_hash: BytesN<32>,
@@ -30,6 +31,7 @@ pub trait LiquidityPoolTrait {
         share_token_decimals: u32,
         swap_fee_bps: i64,
         fee_recipient: Address,
+        max_allowed_slippage_bps: i64,
     ) -> Result<(), ContractError>;
 
     // Deposits token_a and token_b. Also mints pool shares for the "to" Identifier. The amount minted
@@ -82,6 +84,7 @@ pub trait LiquidityPoolTrait {
 
 #[contractimpl]
 impl LiquidityPoolTrait for LiquidityPool {
+    #[allow(clippy::too_many_arguments)]
     fn initialize(
         env: Env,
         token_wasm_hash: BytesN<32>,
@@ -90,6 +93,7 @@ impl LiquidityPoolTrait for LiquidityPool {
         share_token_decimals: u32,
         swap_fee_bps: i64,
         fee_recipient: Address,
+        max_allowed_slippage_bps: i64,
     ) -> Result<(), ContractError> {
         // Token order validation to make sure only one instance of a pool can exist
         if token_a >= token_b {
@@ -118,6 +122,7 @@ impl LiquidityPoolTrait for LiquidityPool {
             pair_type: PairType::Xyk,
             total_fee_bps: validate_fee_bps(&env, swap_fee_bps)?,
             fee_recipient,
+            max_allowed_slippage: max_allowed_slippage_bps,
         };
         save_config(&env, config);
         utils::save_total_shares(&env, 0);
@@ -158,14 +163,15 @@ impl LiquidityPoolTrait for LiquidityPool {
             pool_balance_b,
         )?;
 
+        let config = get_config(&env)?;
+
         assert_slippage_tolerance(
             &env,
             custom_slippage_bps,
             &[amounts.0, amounts.1],
             &[pool_balance_a, pool_balance_b],
+            config.max_allowed_slippage(),
         )?;
-
-        let config = get_config(&env)?;
 
         let token_a_client = token_contract::Client::new(&env, &config.token_a);
         let token_b_client = token_contract::Client::new(&env, &config.token_b);
@@ -407,9 +413,9 @@ fn assert_slippage_tolerance(
     slippage_tolerance: Option<i64>,
     deposits: &[i128; 2],
     pools: &[i128; 2],
+    max_allowed_slippage: Decimal,
 ) -> Result<(), ContractError> {
     let default_slippage = Decimal::percent(1); // Representing 1% as the default slippage tolerance
-    let max_allowed_slippage = Decimal::percent(50); // Representing 50% as the maximum allowed slippage tolerance
 
     let slippage_tolerance = if let Some(slippage_tolerance) = slippage_tolerance {
         Decimal::bps(slippage_tolerance)
@@ -513,24 +519,46 @@ mod tests {
         // Test case that should pass:
         // slippage tolerance of 5000 (0.5 or 50%), deposits of 10 and 20, pools of 30 and 60
         // The price changes fall within the slippage tolerance
-        assert_slippage_tolerance(&env, Some(5_000i64), &[10, 20], &[30, 60]).unwrap();
+        let max_allowed_slippage = 5_000i64;
+        assert_slippage_tolerance(
+            &env,
+            Some(max_allowed_slippage),
+            &[10, 20],
+            &[30, 60],
+            Decimal::bps(max_allowed_slippage),
+        )
+        .unwrap();
     }
 
     #[test]
     fn test_assert_slippage_tolerance_fail_tolerance_too_high() {
         let env = Env::default();
         // Test case that should fail due to slippage tolerance being too high
-        let result =
-            assert_slippage_tolerance(&env, Some(60_000), &[10, 20], &[30, 60]).unwrap_err();
+        let max_allowed_slippage = Decimal::bps(5_000i64);
+        let result = assert_slippage_tolerance(
+            &env,
+            Some(60_000),
+            &[10, 20],
+            &[30, 60],
+            max_allowed_slippage,
+        )
+        .unwrap_err();
         assert_eq!(ContractError::SlippageToleranceExceeded, result);
     }
 
     #[test]
     fn test_assert_slippage_tolerance_fail_slippage_violated() {
         let env = Env::default();
+        let max_allowed_slippage = Decimal::bps(5_000i64);
         // The price changes from 10/15 (0.67) to 40/40 (1.00), violating the 10% slippage tolerance
-        let result =
-            assert_slippage_tolerance(&env, Some(1_000), &[10, 15], &[40, 40]).unwrap_err();
+        let result = assert_slippage_tolerance(
+            &env,
+            Some(1_000),
+            &[10, 15],
+            &[40, 40],
+            max_allowed_slippage,
+        )
+        .unwrap_err();
         assert_eq!(ContractError::SlippageToleranceViolated, result);
     }